Refactor esc listener into hook and use in modal and popover
